🏗️ Reorganize transcoding module structure following colocation principle
- Move VideoTranscoder port and FFmpegVideoTranscoderAdapter to unified transcoding/ folder
- Apply hexagonal architecture colocation principle: "keep files that change together close to each other"
- Create index.ts for clean public API exports
- Update all import paths to use ~/modules/video/transcoding
- Remove empty ports/ and adapters/ directories

This reorganization improves maintainability by grouping related transcoding components that evolve together, following clean architecture best practices.
